+++ telemetry-core/src/main/java/com/newrelic/telemetry/metrics/MetricBuffer.java
@@ -59,7 +59,7 @@ public final class MetricBuffer {
    */
   public void addMetric(Count countMetric) {
     Map<String, Object> attributes = countMetric.getAttributes();
-    ingestWarnings.raiseIngestWarnings(attributes, "Metric");
+    ingestWarnings.raiseIngestWarnings(attributes, countMetric);
     metrics.add(countMetric);
   }
 
@@ -70,7 +70,7 @@ public final class MetricBuffer {
    */
   public void addMetric(Gauge gaugeMetric) {
     Map<String, Object> attributes = gaugeMetric.getAttributes();
-    ingestWarnings.raiseIngestWarnings(attributes, "Metric");
+    ingestWarnings.raiseIngestWarnings(attributes, gaugeMetric);
     metrics.add(gaugeMetric);
   }
 
@@ -81,7 +81,7 @@ public final class MetricBuffer {
    */
   public void addMetric(Summary summaryMetric) {
     Map<String, Object> attributes = summaryMetric.getAttributes();
-    ingestWarnings.raiseIngestWarnings(attributes, "Metric");
+    ingestWarnings.raiseIngestWarnings(attributes, summaryMetric);
     metrics.add(summaryMetric);
   }