+++ telemetry-core/src/test/java/com/newrelic/telemetry/util/IngestWarningsTest.java
@@ -3,8 +3,10 @@ package com.newrelic.telemetry.util;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-import java.util.HashMap;
-import java.util.Map;
+import com.newrelic.telemetry.Attributes;
+import com.newrelic.telemetry.events.Event;
+import com.newrelic.telemetry.logs.Log;
+import com.newrelic.telemetry.metrics.Count;
 import java.util.Random;
 import org.junit.jupiter.api.Test;
 
@@ -13,75 +15,109 @@ class IngestWarningsTest {
   @Test
   void validNumberOfAttributesForEventTest() {
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+    Attributes testAttributes = new Attributes();
     testAttributes.put("test", 1);
     testAttributes.put("name", "bob");
     testAttributes.put("sunny", true);
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Event");
-    verify(testIngestWarnings, never()).eventWarningNumAttributes();
+    long timestamp = 300;
+    Event testEvent = new Event("SampleEvent", testAttributes, timestamp);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testEvent);
+    verify(testIngestWarnings, never()).warningNumAttributes("Event");
   }
 
   @Test
   void invalidNumberOfAttributesForEventTest() {
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+    Attributes testAttributes = new Attributes();
 
     for (int i = 0; i < 300; i++) {
       testAttributes.put("Attribute" + i, i);
     }
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Event");
-    verify(testIngestWarnings).eventWarningNumAttributes();
+    long timestamp = 300;
+    Event testEvent = new Event("SampleEvent", testAttributes, timestamp);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testEvent);
+    verify(testIngestWarnings).warningNumAttributes("Event");
   }
 
   @Test
   void validNumberOfAttributesForMetricTest() {
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+    Attributes testAttributes = new Attributes();
     testAttributes.put("test", 1);
     testAttributes.put("name", "bob");
     testAttributes.put("sunny", true);
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Metric");
-    verify(testIngestWarnings, never()).metricWarningNumAttributes();
+    long testVal = 7;
+    long testStartTimeMs = 100;
+    long testEndTimeMs = 200;
+
+    Count testCount =
+        new Count("TestCount", testVal, testStartTimeMs, testEndTimeMs, testAttributes);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testCount);
+    verify(testIngestWarnings, never()).warningNumAttributes("Metric");
   }
 
   @Test
   void invalidNumberOfAttributesForMetricTest() {
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+    Attributes testAttributes = new Attributes();
     for (int i = 0; i < 110; i++) {
       testAttributes.put("Attribute" + i, i);
     }
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Metric");
-    verify(testIngestWarnings).metricWarningNumAttributes();
+    long testVal = 7;
+    long testStartTimeMs = 100;
+    long testEndTimeMs = 200;
+
+    Count testCount =
+        new Count("TestCount", testVal, testStartTimeMs, testEndTimeMs, testAttributes);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testCount);
+    verify(testIngestWarnings).warningNumAttributes("Metric");
   }
 
   @Test
   void validNumberOfAttributesForLogTest() {
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+    Attributes testAttributes = new Attributes();
     testAttributes.put("test", 1);
     testAttributes.put("name", "bob");
     testAttributes.put("sunny", true);
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Log");
-    verify(testIngestWarnings, never()).logWarningNumAttributes();
+    Log testLog =
+        Log.builder()
+            .attributes(testAttributes)
+            .message("Processing Information")
+            .level("DEBUG")
+            .build();
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testLog);
+    verify(testIngestWarnings, never()).warningNumAttributes("Log");
   }
 
   @Test
   void invalidNumberOfAttributesForLogTest() {
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+    Attributes testAttributes = new Attributes();
 
     for (int i = 0; i < 300; i++) {
       testAttributes.put("Attribute" + i, i);
     }
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Log");
-    verify(testIngestWarnings).logWarningNumAttributes();
+    Log testLog =
+        Log.builder()
+            .attributes(testAttributes)
+            .message("Processing Information")
+            .level("DEBUG")
+            .build();
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testLog);
+    verify(testIngestWarnings).warningNumAttributes("Log");
   }
 
   // add # of attributes test(s) for logs
@@ -89,12 +125,16 @@ class IngestWarningsTest {
   @Test
   void validAttributeNamesTest() {
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+
+    Attributes testAttributes = new Attributes();
     testAttributes.put("test", 1);
     testAttributes.put("name", "bob");
     testAttributes.put("sunny", true);
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Event");
+    long timestamp = 300;
+    Event testEvent = new Event("SampleEvent", testAttributes, timestamp);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testEvent);
     verify(testIngestWarnings, never()).attributeNameWarning("test");
   }
 
@@ -115,25 +155,32 @@ class IngestWarningsTest {
     String longAttrName = sb.toString();
 
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+
+    Attributes testAttributes = new Attributes();
     testAttributes.put(longAttrName, 1);
     testAttributes.put("name", "bob");
     testAttributes.put("sunny", true);
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Event");
+    long timestamp = 300;
+    Event testEvent = new Event("SampleEvent", testAttributes, timestamp);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testEvent);
     verify(testIngestWarnings).attributeNameWarning(longAttrName);
   }
 
   @Test
   void validAttributeValuesTest() {
-
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+
+    Attributes testAttributes = new Attributes();
     testAttributes.put("test", 1);
     testAttributes.put("name", "bob");
     testAttributes.put("sunny", true);
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Event");
+    long timestamp = 300;
+    Event testEvent = new Event("SampleEvent", testAttributes, timestamp);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testEvent);
     verify(testIngestWarnings, never()).attributeValueWarning("test");
   }
 
@@ -153,12 +200,17 @@ class IngestWarningsTest {
     String longAttrValue = sb.toString();
 
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+
+    Attributes testAttributes = new Attributes();
     testAttributes.put("test", longAttrValue);
     testAttributes.put("name", "bob");
     testAttributes.put("sunny", true);
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Event");
+    long timestamp = 300;
+    Event testEvent = new Event("SampleEvent", testAttributes, timestamp);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testEvent);
+
     verify(testIngestWarnings).attributeValueWarning(longAttrValue);
   }
 
@@ -178,12 +230,16 @@ class IngestWarningsTest {
     String longAttrValue = sb.toString();
 
     IngestWarnings testIngestWarnings = spy(new IngestWarnings());
-    Map<String, Object> testAttributes = new HashMap<>();
+
+    Attributes testAttributes = new Attributes();
     testAttributes.put(longAttrValue, longAttrValue);
     testAttributes.put("name", "bob");
     testAttributes.put("sunny", true);
 
-    testIngestWarnings.raiseIngestWarnings(testAttributes, "Event");
+    long timestamp = 300;
+    Event testEvent = new Event("SampleEvent", testAttributes, timestamp);
+
+    testIngestWarnings.raiseIngestWarnings(testAttributes.asMap(), testEvent);
 
     verify(testIngestWarnings).attributeNameWarning(longAttrValue);
     verify(testIngestWarnings).attributeValueWarning(longAttrValue);