+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.13.0] - TBD 
 - EU Endpoint Support added by [updating the SenderConfigurationBuilder API](https://github.com/newrelic/newrelic-telemetry-sdk-java/pull/276). 
 - Includes endpoints that send Metric, Event, Log, and Span data to New Relic One.
+- Added warnings to let users know if data exceeds some Ingest API limits. 
 
 ## [0.12.0] - 2021-02-19
 - Accepts a New Relic APM license key as an alternative to an Insights Insert API key.